requirements.txt
oeispy
discord.py